requirements.txt
pandas
python-vlc
Pillow
